requirements.txt
streamlit==1.28.0
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
transformers==4.33.0
diffusers==0.21.4
accelerate==0.21.0
safety-checker==0.0.5
python-dotenv==1.0.0
numpy==1.24.3
opencv-python==4.8.0.74
requests==2.31.0
